Integrate API in StartView
Implements uploading files to the backend via the upload button. We
should probably implement some kind of file validation in the backend,
i.e validate that it is valid WASM, exports the functions we expect etc.

Also implements fetching the bots on load and populating the dropdowns
with the values.