Implement an optimal tic-tac-toe bot
This bot utilizes an algorithm called Negamax (which is just a nicer
version of minimax). In essense it looks ahead as many moves as possible
and assumes it's opponent is playing optimally and from there determines
an optimal move.

Having a perfect winning strategy is impossible in tic-tac-toe so this
bot can only guarantee it will not lose.

Written in Rust!