Implement heuristic-based bot (C++)
Adds a bot written in C++. Unlike the optimal bot, this one is based on
a heuristic, assigning an estimated score on how good a board is. From
all available moves it then picks the one with the highest score.

I belive this bot is also unbeatable, i.e cannot lose.